Moved tarot logic away from teekkari.py to the tarot class
